requirements.txt
asgiref==3.9.1
Django==5.2.5
djangorestframework==3.16.1
et_xmlfile==2.0.0
greenlet==3.2.3
openpyxl==3.1.5
playwright==1.54.0
pyee==13.0.0
sqlparse==0.5.3
typing_extensions==4.14.1
